% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfCallGeno.R
\name{cfCallGeno}
\alias{cfCallGeno}
\title{Calculate maximum likelihood genotype}
\usage{
cfCallGeno(mar, dep, ff)
}
\arguments{
\item{mar}{Integer of length 1, minor allele read count}

\item{dep}{Integer of length 1, total read depth}

\item{ff}{Numeric of length 1, fetal fraction}
}
\value{
Posterior probability for the maximum likelihood genotype
}
\description{
Calculate maximum likelihood genotype
}
\details{
Add later.
}
