% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfEstFF.R
\name{cfEstFF}
\alias{cfEstFF}
\title{Estimate fetal fraction}
\usage{
cfEstFF(marVec, depVec, gtpVec)
}
\arguments{
\item{marVec}{Integer, minor allele read counts}

\item{depVec}{Integer, total read depths}

\item{gtpVec}{Character, genotype pairs, e.g. "ABab"}
}
\description{
Estimate fetal fraction
}
