% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalBayes.R
\name{empiricalBayesFuncs}
\alias{empiricalBayesFuncs}
\alias{cfUpdateEmpGeno}
\alias{cfUpdateEmpPMAR}
\alias{cfUpdateEmpGenoFreq}
\alias{cfEmpGeno}
\title{Functions for performing empirical Bayesian calculations}
\usage{
cfUpdateEmpGeno(mar, dep, empPMAR, empGenoFreq)

cfUpdateEmpPMAR(marVec, depVec, empGeno, ffPrior, ffPriorWt)

cfUpdateEmpGenoFreq(empGeno, qPrior, qPriorWt)

cfEmpGeno(
  marVec,
  depVec,
  qPrior,
  ffPrior,
  nchng = 5,
  its = 50,
  ffPriorWt = length(marVec)/5,
  qPriorWt = 1000,
  pltTrain = FALSE
)
}
\arguments{
\item{mar}{Integer of length 1, minor allele read count}

\item{dep}{Integer of length 1, total read depth}

\item{empPMAR}{Named numeric, empirical estimates of expected PMAR values}

\item{empGenoFreq}{Named numeric, empirical estimates of genotype pair
frequencies}

\item{marVec}{Integer, minor allele read counts}

\item{depVec}{Integer, total read depths}

\item{empGeno}{Character, empirical genotype calls}

\item{ffPrior}{Numeric of length 1, fetal fraction prior}

\item{ffPriorWt}{Numeric of length 1, the weight given to the fetal fraction
prior}

\item{qPrior}{Numeric of length 1, average minor allele frequency prior}

\item{qPriorWt}{Numeric of length 1, the weight given to the q prior}

\item{pltTrain}{Logical of length 1, plot genotypes at each training
iteration when TRUE}
}
\description{
Collection of functions for performing empirical Bayesian
calculations
}
\details{
ffPriorWt can be thought of as the number of observations supportint the
fetal fraction prior
}
\section{Functions}{
\itemize{
\item \code{cfUpdateEmpGeno}: Update the emperic genotype call

\item \code{cfUpdateEmpPMAR}: Update the emperic PMAR values

\item \code{cfUpdateEmpGenoFreq}: Update the emperic genotype frequencies

\item \code{cfEmpGeno}: Make empiric genotype calls
}}

