% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedValues.R
\name{expectedValues}
\alias{expectedValues}
\alias{cfExpPMAR}
\alias{cfExpSdPMAR}
\alias{cfExpUnqHet}
\alias{cfExpGenoFreq}
\title{Calculate expected values}
\usage{
cfExpPMAR(ffVec, seqError = 0.003, clpsList = TRUE)

cfExpSdPMAR(ffVec, dep, seqError = 0.003, clpsList = TRUE)

cfExpUnqHet(qVec)

cfExpGenoFreq(qVec, clpsList = TRUE)
}
\arguments{
\item{ffVec}{Numeric, fetal fractions}

\item{clpsList}{List items will be collapsed to a vector when possible}

\item{dep}{Integer of length 1, total read depth}

\item{qVec}{Numeric, minor allele frequencies}
}
\description{
Collection of functions for calculating various expected values
}
\details{
Default seqError of 0.003 taken from PMID30026539 (mean + 1SD to over-
estimate).
}
\section{Functions}{
\itemize{
\item \code{cfExpPMAR}: Expected proportion of minor allele reads

\item \code{cfExpSdPMAR}: Expected standard deviation on proportion of minor
allele reads

\item \code{cfExpUnqHet}: Expected proportion of sites with unique fetal
alleles

\item \code{cfExpGenoFreq}: Expected proportion of genotype pairs
}}

