% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgParams.R
\name{pkgParams}
\alias{pkgParams}
\title{Common package parameters}
\arguments{
\item{ff}{Numeric of length 1, fetal fraction}

\item{ffVec}{Numeric, fetal fractions}

\item{dep}{Integer of length 1, total read depth}

\item{depVec}{Integer, total read depths}

\item{q}{Numeric of length 1, minor allele frequency}

\item{qVec}{Numeric, minor allele frequencies}

\item{clpsList}{List items will be collapsed to a vector when possible}

\item{mar}{Integer of length 1, minor allele read count}

\item{marVec}{Integer, minor allele read counts}

\item{gtpVec}{Character, genotype pairs, e.g. "ABab"}
}
\description{
Common package parameters
}
